Rename to PackageDetailsFormContainer
