+++ components/admin/PackageEdit/PackageDetailsForm/PackageDetailsFormContainer/mocks.js
